Organize MCP related code
